crud de usuarios e produtos pronto

+++ src/main/java/com/projeto/CafeAromaESabor/Repository/ProdutoRepository.java
@@ -3,5 +3,5 @@ package com.projeto.CafeAromaESabor.Repository;
 import com.projeto.CafeAromaESabor.Model.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 }
