+++ src/main/java/com/projeto/CafeAromaESabor/Repository/UserRepository.java
@@ -3,6 +3,8 @@ package com.projeto.CafeAromaESabor.Repository;
 import com.projeto.CafeAromaESabor.Model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer>
+import java.util.List;
+
+public interface UserRepository extends JpaRepository<User, Long>
 {
 }
